Implement KEDA and HPA integration for autoscaler functionality, enhancing minReplica resolution and updating PDB handling
